added visualations.py file & some graph formatting
