+++ helper_methods.py
@@ -352,75 +352,3 @@ def calculate_profit_no_stop_loss(stock_1, stock_2, data, train_data):
         position_stock_2 = next_day_position_stock_2
     # Close positions at stop threshold or after time period ends
     return (balance, daily_profits)
-
-# Visualizations
-
-def plot_stock_pair(series_one, label_one, series_two, label_two):
-    _fig, ax = plt.subplots(figsize=(8, 4))
-    ax.plot(series_one, label=label_one)
-    ax.plot(series_two, label=label_two)
-    ax.legend()
-    plt.show()
-
-def plot_stock_one_adjusted_pair(series_one, label_one, series_two, label_two):
-    series_one_adjusted = series_one / \
-        (np.mean(series_one) / np.mean(series_two))
-    plot_stock_pair(series_one_adjusted, label_one, series_two, label_two)
-
-def plot_stock_pair_ratio(series_one, label_one, series_two, label_two):
-    # Use ratios to calculate spread
-    ratio = series_one / series_two
-
-    # Normalize the series
-    z_score = (ratio - ratio.mean()) / ratio.std()
-
-    fig, ax = plt.subplots(figsize=(8, 4))
-    ax.plot(z_score, label=f'Ratio ({label_one} / {label_two})')
-    plt.axhline(z_score.mean(), color='black')
-
-    # ~80% of the data lies within our threshold
-    plt.axhline(trade_threshold(z_score), color = 'blue')
-    plt.axhline(-trade_threshold(z_score), color = 'blue')
-
-    ax.legend()
-    plt.show()
-
-def plot_mvg_avg_ratio(price_ratio_series, train_data):
-    # Plot new graph based off moving avg
-    fig, ax = plt.subplots(figsize=(8, 4))
-    ax.plot(mvg_avg_z_score(price_ratio_series), label='mvg_avg ratio z_score')
-    plt.axhline(0, color='black')
-    plt.axhline(trade_threshold(mvg_avg_z_score(train_data)), color = 'blue')
-    plt.axhline(-trade_threshold(mvg_avg_z_score(train_data)), color = 'blue')
-    plt.legend()
-    plt.show()
-
-def plot_trade_signals(ratio_series, train_data):
-    fig, ax = plt.subplots(figsize=(8, 4))
-    ax.plot(ratio_series, label='label')
-    ax.scatter(ratio_series.index[trade_signals(ratio_series, train_data) == 1], ratio_series[trade_signals(ratio_series, train_data) == 1], color='green', marker='^', label='Buy (1)')
-    ax.scatter(ratio_series.index[trade_signals(ratio_series, train_data) == -1], ratio_series[trade_signals(ratio_series, train_data) == -1], color='red', marker='v', label='Sell (-1)')
-    plt.show()
-
-def plot_trade_signals_no_stop_loss(ratio_series, train_data):
-    fig, ax = plt.subplots(figsize=(8, 4))
-    ax.plot(ratio_series, label='label')
-    ax.scatter(ratio_series.index[trade_signals_no_stop_loss(ratio_series, train_data) == 1], ratio_series[trade_signals_no_stop_loss(ratio_series, train_data) == 1], color='green', marker='^', label='Buy (1)')
-    ax.scatter(ratio_series.index[trade_signals_no_stop_loss(ratio_series, train_data) == -1], ratio_series[trade_signals_no_stop_loss(ratio_series, train_data) == -1], color='red', marker='v', label='Sell (-1)')
-    plt.show()
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
